+++ application/command-composition/src/main/java/org/occurrent/application/composition/command/ListCommandComposition.java
@@ -20,7 +20,7 @@ package org.occurrent.application.composition.command;
 
 import org.occurrent.application.composition.command.internal.CreateListFromVarArgs;
 import org.occurrent.application.composition.command.internal.SequentialFunctionComposer;
-import org.occurrent.application.composition.command.partial.PartialListCommandApplication;
+import org.occurrent.application.composition.command.partial.PartialFunctionApplication;
 
 import java.util.List;
 import java.util.function.Function;
@@ -44,7 +44,7 @@ import java.util.stream.Stream;
  * <br>
  * <br>
  * Note that in most cases the domain function will not have the form {@code Function<List<T>, List<T>>}. You can then
- * use {@link PartialListCommandApplication} to create partially applied functions that you can then compose.
+ * use {@link PartialFunctionApplication} to create partially applied functions that you can then compose.
  */
 public class ListCommandComposition {